extracted Account component to separate directory
